Start to integrate Chronos.
Extract a DateFormatTrait as soon there will be an Cake\I18n\Date class
to provide Time improvements for Date columns.
